+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.1] - Customized imageUrl
+
+* Gravatar imageUrl with custom attributes.  
+
 ## [1.0.0] - Initial release.
 
 * Gravatar imageUrl and profileUrl.